requirements.txt
python-telegram-bot>=20.0
speedtest-cli>=2.1.3
psutil>=5.9.0
pytz>=2023.3
apscheduler==3.9.1  # Pinned to avoid timezone issues